backend/django_backend/requirements.txt
django>=5.0,<5.1
psycopg[binary]
pika==1.3.2
django-timescaledb
djangorestframework
pytz>=2023.3
supabase